Move processmessagechunk file into a common chat function
